Improve naming of domain objects
