+++ src/domain/local_branch_name_test.go
@@ -36,21 +36,3 @@ func TestLocalBranchName(t *testing.T) {
 		assert.Equal(t, want, branch.RemoteName())
 	})
 }
-
-func TestLocalBranchNames(t *testing.T) {
-	t.Parallel()
-	t.Run("NewLocalBranchNames and Strings", func(t *testing.T) {
-		t.Parallel()
-		branches := domain.NewLocalBranchNames("one", "two", "three")
-		want := []string{"one", "two", "three"}
-		assert.Equal(t, want, branches.Strings())
-	})
-
-	t.Run("Sort", func(t *testing.T) {
-		t.Parallel()
-		branches := domain.NewLocalBranchNames("one", "two", "three")
-		want := []string{"one", "three", "two"}
-		branches.Sort()
-		assert.Equal(t, want, branches.Strings())
-	})
-}
