+++ src/domain/local_branch_names_test.go
@@ -1,0 +1,26 @@
+package domain_test
+
+import (
+	"testing"
+
+	"github.com/git-town/git-town/v9/src/domain"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestLocalBranchNames(t *testing.T) {
+	t.Parallel()
+	t.Run("NewLocalBranchNames and Strings", func(t *testing.T) {
+		t.Parallel()
+		branches := domain.NewLocalBranchNames("one", "two", "three")
+		want := []string{"one", "two", "three"}
+		assert.Equal(t, want, branches.Strings())
+	})
+
+	t.Run("Sort", func(t *testing.T) {
+		t.Parallel()
+		branches := domain.NewLocalBranchNames("one", "two", "three")
+		want := []string{"one", "three", "two"}
+		branches.Sort()
+		assert.Equal(t, want, branches.Strings())
+	})
+}
